requirements.txt

pyrogram==2.0.106
tgcrypto
flask
python-dotenv
requests
